requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv
PyMuPDF
python-multipart
google-genai
psycopg2-binary
alembic

